+++ apps/bare-expo/modules/benchmarking/src/index.ts
@@ -3,3 +3,6 @@ import ExpoModule from './BenchmarkingExpoModule';
 import TurboModule from './NativeBenchmarkingTurboModule';
 
 export { TurboModule, ExpoModule, BridgeModule };
+export { default as BenchmarkView } from './BenchmarkView';
+export { default as ViewPropsBenchmarkScreen } from './ViewPropsBenchmarkScreen';
+export type { ViewPropsBenchmark } from './BenchmarkingExpoModule';